Move navigation to shared.
